Remove test helper
